+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/usecase/RequestSwapQuoteUseCase.kt
@@ -53,7 +53,7 @@ class RequestSwapQuoteUseCase(
                 swapRepository.requestExactInputQuote(
                     amount = amount,
                     address = address,
-                    refundAddress = newAddress.address,
+                    refundAddress = newAddress,
                     destinationAsset = selectedAsset,
                     slippage = slippage
                 )
@@ -76,7 +76,7 @@ class RequestSwapQuoteUseCase(
                 swapRepository.requestExactOutputQuote(
                     amount = amount,
                     address = address,
-                    refundAddress = newAddress.address,
+                    refundAddress = newAddress,
                     destinationAsset = selectedAsset,
                     slippage = slippage
                 )
@@ -100,7 +100,7 @@ class RequestSwapQuoteUseCase(
                     .requestFlexInputIntoZec(
                         amount = amount,
                         refundAddress = refundAddress,
-                        destinationAddress = newAddress.address,
+                        destinationAddress = newAddress,
                         originAsset = selectedAsset,
                         slippage = slippage
                     )